Remove insecure package
Previously, we were able to make db calls from the client using
Jobs.insert(), which is insecure.  This commit removes the insecure
package (which allows this prototyping functionality) and replaces it
with Meteor methods and calls.  The methods are defined server-side and
interface with the db, but can also do things like handling validation
(ensuring the user is logged in, in our case).  The call allows the
defined method to be triggered, passing in data if necessary.